Rename Test to IntegrationTest and moved sample arrays in test methods
